Split userland service providers from broker main

Copilot-Session: 239ec5f7-870a-4259-bcae-4ca85fb913a0
