+++ litebox_broker_userland/src/main.rs
@@ -1,15 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-use std::collections::VecDeque;
 use std::error::Error;
 use std::ffi::{OsStr, OsString};
-use std::io::{Error as IoError, ErrorKind, Read as _, Result as IoResult};
+use std::io::{Error as IoError, ErrorKind, Result as IoResult};
 use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::process::{Child, Command};
 use std::str::FromStr;
-use std::sync::mpsc::{Receiver, RecvTimeoutError, SyncSender, TrySendError, sync_channel};
+use std::sync::mpsc::{Receiver, SyncSender, TrySendError, sync_channel};
 use std::sync::{
     Arc, Mutex,
     atomic::{AtomicBool, Ordering},
@@ -17,17 +16,14 @@ use std::sync::{
 use std::time::{Duration, Instant};
 
 use clap::Parser;
-use litebox_broker_core::random::{RandomProvider, RandomProviderError};
-use litebox_broker_core::stdio::{StdioProvider, StdioProviderError};
 use litebox_broker_core::{
-    AssociationCancellation, BrokerCore, CallerCredential, DestinationPortRange, DestinationRule,
-    Ipv4Cidr, SocketPolicy, SocketPolicyError,
+    BrokerCore, CallerCredential, DestinationPortRange, DestinationRule, Ipv4Cidr, SocketPolicy,
+    SocketPolicyError,
 };
 use litebox_broker_host::{BrokerHostAssociation, ConnectionTermination, setup_connection};
 use litebox_broker_protocol::message::{BrokerRequest, BrokerResponse};
 use litebox_broker_protocol::shared_buffer::SHARED_BUFFER_LAYOUT;
 use litebox_broker_protocol::socket::{Ipv4Address, Port};
-use litebox_broker_protocol::stdio::{MAX_STDIO_TRANSFER_SIZE, StdioOutputStream};
 use litebox_broker_transport::channel::{HostNotificationChannel, HostReceive, HostSetupChannel};
 use litebox_broker_transport::control_ring::ControlRing;
 use litebox_broker_transport::shared_memory::{ControlRingMemory, SharedBufferPool, SharedMemory};
@@ -36,205 +32,17 @@ use litebox_broker_userland::readiness::ReadinessPublisherRuntime;
 
 #[cfg(target_os = "linux")]
 mod linux;
+mod random;
+mod stdio;
 #[cfg(all(windows, target_arch = "x86_64"))]
 mod windows;
 
 const SETUP_TIMEOUT: Duration = Duration::from_secs(5);
 const ACCEPT_RETRY_DELAY: Duration = Duration::from_millis(10);
-const STDIO_CANCEL_POLL_INTERVAL: Duration = Duration::from_millis(50);
-const STDIO_WRITE_RETRY_DELAY: Duration = Duration::from_millis(1);
 const REQUEST_QUEUE_CAPACITY: usize = 64;
 const REQUEST_QUEUE_RETRY_DELAY: Duration = Duration::from_millis(1);
 const REQUEST_QUEUE_STALL_TIMEOUT: Duration = Duration::from_secs(5);
 const WORKER_COUNT: usize = 8;
-
-struct UserlandRandomProvider;
-
-impl RandomProvider for UserlandRandomProvider {
-    fn fill(&self, output: &mut [u8]) -> Result<(), RandomProviderError> {
-        getrandom::fill(output).map_err(|_| RandomProviderError)
-    }
-}
-
-/// Routes standard I/O for the broker's single child runner through inherited
-/// streams.
-///
-/// A broker serving multiple runners will need association-specific stream
-/// endpoints instead of sharing process-wide standard streams.
-struct UserlandStdioProvider {
-    stdin: Mutex<UserlandStdin>,
-    stdout: SyncSender<StdioWriteRequest>,
-    stderr: SyncSender<StdioWriteRequest>,
-}
-
-struct UserlandStdin {
-    receiver: Receiver<StdinReadResult>,
-    buffered: VecDeque<u8>,
-    eof: bool,
-}
-
-enum StdinReadResult {
-    Data(Vec<u8>),
-    Eof,
-    Error(StdioProviderError),
-}
-
-struct StdioWriteRequest {
-    input: Vec<u8>,
-    completion: SyncSender<Result<usize, StdioProviderError>>,
-}
-
-impl UserlandStdioProvider {
-    fn new() -> IoResult<Self> {
-        let (stdin_sender, stdin_receiver) = sync_channel(1);
-        std::thread::Builder::new()
-            .name("litebox-broker-stdin".to_owned())
-            .spawn(move || pump_stdin(stdin_sender))?;
-        let (stdout, stdout_receiver) = sync_channel(WORKER_COUNT);
-        std::thread::Builder::new()
-            .name("litebox-broker-stdout".to_owned())
-            .spawn(move || pump_stdout(stdout_receiver))?;
-        let (stderr, stderr_receiver) = sync_channel(WORKER_COUNT);
-        std::thread::Builder::new()
-            .name("litebox-broker-stderr".to_owned())
-            .spawn(move || pump_stderr(stderr_receiver))?;
-        Ok(Self {
-            stdin: Mutex::new(UserlandStdin {
-                receiver: stdin_receiver,
-                buffered: VecDeque::new(),
-                eof: false,
-            }),
-            stdout,
-            stderr,
-        })
-    }
-}
-
-impl StdioProvider for UserlandStdioProvider {
-    fn read(
-        &self,
-        cancellation: &AssociationCancellation,
-        output: &mut [u8],
-    ) -> Result<usize, StdioProviderError> {
-        if output.is_empty() {
-            return Ok(0);
-        }
-        let mut stdin = self.stdin.lock().map_err(|_| StdioProviderError::Failed)?;
-        loop {
-            if cancellation.is_cancelled() {
-                return Err(StdioProviderError::Closed);
-            }
-            if !stdin.buffered.is_empty() {
-                let read = output.len().min(stdin.buffered.len());
-                for destination in &mut output[..read] {
-                    *destination = stdin
-                        .buffered
-                        .pop_front()
-                        .expect("buffered stdin length was checked");
-                }
-                return Ok(read);
-            }
-            if stdin.eof {
-                return Ok(0);
-            }
-            match stdin.receiver.recv_timeout(STDIO_CANCEL_POLL_INTERVAL) {
-                Ok(StdinReadResult::Data(data)) => stdin.buffered.extend(data),
-                Ok(StdinReadResult::Eof) => stdin.eof = true,
-                Ok(StdinReadResult::Error(error)) => return Err(error),
-                Err(RecvTimeoutError::Timeout) => {}
-                Err(RecvTimeoutError::Disconnected) => return Err(StdioProviderError::Failed),
-            }
-        }
-    }
-
-    fn write(
-        &self,
-        cancellation: &AssociationCancellation,
-        stream: StdioOutputStream,
-        input: &[u8],
-    ) -> Result<usize, StdioProviderError> {
-        let sender = match stream {
-            StdioOutputStream::Stdout => &self.stdout,
-            StdioOutputStream::Stderr => &self.stderr,
-        };
-        let (completion, completed) = sync_channel(1);
-        let mut request = StdioWriteRequest {
-            input: input.to_vec(),
-            completion,
-        };
-        loop {
-            if cancellation.is_cancelled() {
-                return Err(StdioProviderError::Closed);
-            }
-            match sender.try_send(request) {
-                Ok(()) => break,
-                Err(TrySendError::Full(pending)) => request = pending,
-                Err(TrySendError::Disconnected(_)) => {
-                    return Err(StdioProviderError::Failed);
-                }
-            }
-            std::thread::sleep(STDIO_WRITE_RETRY_DELAY);
-        }
-        loop {
-            match completed.recv_timeout(STDIO_CANCEL_POLL_INTERVAL) {
-                Ok(result) => return result,
-                Err(RecvTimeoutError::Timeout) if cancellation.is_cancelled() => {
-                    return Err(StdioProviderError::Closed);
-                }
-                Err(RecvTimeoutError::Timeout) => {}
-                Err(RecvTimeoutError::Disconnected) => {
-                    return Err(StdioProviderError::Failed);
-                }
-            }
-        }
-    }
-}
-
-fn pump_stdin(sender: SyncSender<StdinReadResult>) {
-    let stdin = std::io::stdin();
-    let mut stdin = stdin.lock();
-    let mut buffer = vec![0; MAX_STDIO_TRANSFER_SIZE as usize];
-    loop {
-        let (result, finished) = match stdin.read(&mut buffer) {
-            Ok(0) => (StdinReadResult::Eof, true),
-            Ok(read) => (StdinReadResult::Data(buffer[..read].to_vec()), false),
-            Err(error) if error.kind() == ErrorKind::Interrupted => continue,
-            Err(error) => (StdinReadResult::Error(map_stdio_error(error)), true),
-        };
-        if sender.send(result).is_err() || finished {
-            return;
-        }
-    }
-}
-
-fn pump_stdout(receiver: Receiver<StdioWriteRequest>) {
-    pump_output(std::io::stdout(), receiver);
-}
-
-fn pump_stderr(receiver: Receiver<StdioWriteRequest>) {
-    pump_output(std::io::stderr(), receiver);
-}
-
-fn pump_output(mut output: impl std::io::Write, receiver: Receiver<StdioWriteRequest>) {
-    while let Ok(request) = receiver.recv() {
-        let result = write_and_flush(&mut output, &request.input).map_err(map_stdio_error);
-        let _ = request.completion.send(result);
-    }
-}
-
-fn map_stdio_error(error: IoError) -> StdioProviderError {
-    if error.kind() == ErrorKind::BrokenPipe {
-        StdioProviderError::Closed
-    } else {
-        StdioProviderError::Failed
-    }
-}
-
-fn write_and_flush(mut output: impl std::io::Write, input: &[u8]) -> IoResult<usize> {
-    let written = output.write(input)?;
-    output.flush()?;
-    Ok(written)
-}
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 struct AllowedDestination {
@@ -818,39 +626,6 @@ fn main() {}
 mod cli_tests {
     use super::*;
 
-    #[derive(Default)]
-    struct RecordingOutput {
-        bytes: Arc<Mutex<Vec<u8>>>,
-        flushes: Arc<Mutex<usize>>,
-    }
-
-    impl std::io::Write for RecordingOutput {
-        fn write(&mut self, input: &[u8]) -> IoResult<usize> {
-            let written = input.len().min(3);
-            self.bytes.lock().unwrap().extend(&input[..written]);
-            Ok(written)
-        }
-
-        fn flush(&mut self) -> IoResult<()> {
-            *self.flushes.lock().unwrap() += 1;
-            Ok(())
-        }
-    }
-
-    fn test_stdio_provider(receiver: Receiver<StdinReadResult>) -> UserlandStdioProvider {
-        let (stdout, _stdout_receiver) = sync_channel(1);
-        let (stderr, _stderr_receiver) = sync_channel(1);
-        UserlandStdioProvider {
-            stdin: Mutex::new(UserlandStdin {
-                receiver,
-                buffered: VecDeque::new(),
-                eof: false,
-            }),
-            stdout,
-            stderr,
-        }
-    }
-
     #[test]
     fn cli_accepts_tcp_and_udp_destination_arguments() {
         let args = CliArgs::try_parse_from([
@@ -918,63 +693,6 @@ mod cli_tests {
             DestinationRule::new(CallerCredential::HostGuaranteed, udp.destination, udp.ports,)
         );
     }
-
-    #[test]
-    fn stdio_provider_preserves_partial_reads_and_eof() {
-        let (sender, receiver) = sync_channel(2);
-        sender
-            .send(StdinReadResult::Data(b"input".to_vec()))
-            .unwrap();
-        sender.send(StdinReadResult::Eof).unwrap();
-        let provider = test_stdio_provider(receiver);
-        let cancellation = AssociationCancellation::default();
-
-        let mut first = [0xa5; 3];
-        assert_eq!(provider.read(&cancellation, &mut first), Ok(3));
-        assert_eq!(&first, b"inp");
-
-        let mut second = [0xa5; 4];
-        assert_eq!(provider.read(&cancellation, &mut second), Ok(2));
-        assert_eq!(&second, b"ut\xa5\xa5");
-        assert_eq!(provider.read(&cancellation, &mut second), Ok(0));
-    }
-
-    #[test]
-    fn stdio_provider_waits_for_output_write_and_flush() {
-        let bytes = Arc::new(Mutex::new(Vec::new()));
-        let flushes = Arc::new(Mutex::new(0));
-        let (stdout, stdout_receiver) = sync_channel(1);
-        let output = RecordingOutput {
-            bytes: Arc::clone(&bytes),
-            flushes: Arc::clone(&flushes),
-        };
-        let writer = std::thread::spawn(move || pump_output(output, stdout_receiver));
-        let (_stdin_sender, stdin_receiver) = sync_channel(1);
-        let (stderr, _stderr_receiver) = sync_channel(1);
-        let provider = UserlandStdioProvider {
-            stdin: Mutex::new(UserlandStdin {
-                receiver: stdin_receiver,
-                buffered: VecDeque::new(),
-                eof: false,
-            }),
-            stdout,
-            stderr,
-        };
-
-        assert_eq!(
-            provider.write(
-                &AssociationCancellation::default(),
-                StdioOutputStream::Stdout,
-                b"prompt",
-            ),
-            Ok(3)
-        );
-
-        drop(provider);
-        writer.join().unwrap();
-        assert_eq!(*bytes.lock().unwrap(), b"pro");
-        assert_eq!(*flushes.lock().unwrap(), 1);
-    }
 }
 
 #[cfg(all(test, target_os = "linux"))]
@@ -988,7 +706,7 @@ mod tests {
 
     use litebox_broker_core::socket::UnsupportedSocketProvider;
     use litebox_broker_core::stdio::{StdioProvider, StdioProviderError, UnsupportedStdioProvider};
-    use litebox_broker_core::{BrokerCore, ObjectRights, PolicyEngine};
+    use litebox_broker_core::{AssociationCancellation, BrokerCore, ObjectRights, PolicyEngine};
     use litebox_broker_host::setup_connection;
     use litebox_broker_protocol::BROKER_PROTOCOL_VERSION;
     use litebox_broker_protocol::RequestId;
@@ -999,6 +717,7 @@ mod tests {
         SHARED_BUFFER_LAYOUT, SHARED_BUFFER_POOL_SIZE, SharedBufferDescriptor,
         SharedBufferSlotIndex,
     };
+    use litebox_broker_protocol::stdio::StdioOutputStream;
     use litebox_broker_transport::channel::{
         HostNotificationChannel, HostReceive, HostSetupChannel, LocalSetupChannel,
     };
@@ -1013,6 +732,8 @@ mod tests {
     };
 
     use super::*;
+
+    const TEST_CANCELLATION_POLL_INTERVAL: Duration = Duration::from_millis(10);
 
     /// One live host association: the endpoints teardown acts on, and the rest
     /// held open so the association stays up for the duration of a test.
@@ -1124,7 +845,7 @@ mod tests {
             let broker = BrokerCore::new(
                 PolicyEngine::with_host_guaranteed_rights(ObjectRights::all()),
                 Arc::new(UnsupportedSocketProvider),
-                Arc::new(UserlandRandomProvider),
+                Arc::new(random::UserlandRandomProvider),
                 stdio_provider,
             )
             .unwrap();
@@ -1180,7 +901,7 @@ mod tests {
         ) -> Result<usize, StdioProviderError> {
             self.started.send(()).unwrap();
             while !cancellation.is_cancelled() {
-                std::thread::sleep(STDIO_CANCEL_POLL_INTERVAL);
+                std::thread::sleep(TEST_CANCELLATION_POLL_INTERVAL);
             }
             Err(StdioProviderError::Closed)
         }
@@ -1216,7 +937,7 @@ mod tests {
         ) -> Result<usize, StdioProviderError> {
             self.started.send(()).unwrap();
             while !cancellation.is_cancelled() {
-                std::thread::sleep(STDIO_CANCEL_POLL_INTERVAL);
+                std::thread::sleep(TEST_CANCELLATION_POLL_INTERVAL);
             }
             Err(StdioProviderError::Closed)
         }
